requirements.txt
Pyrogram==2.0.106
TgCrypto==1.2.5
Flask==3.0.3
Flask-Cors==4.0.0
Werkzeug==3.0.2
gunicorn==22.0.0
requests==2.31.0
python-dotenv==1.0.1
tenacity==8.2.3
